user-pages: new login page
Replace the old log in page with a simpler one. Do not display login and
registration form on the same page anymore, and instead just give a link
at the bottom to switch between the two of them. This is in preparation
for google oauth, as adding all the oauth stuff to the old page would
have made it incredibly cluttered.

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>